# packages/contracts/openapi/taxmind-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_official_document_api_v1_knowledge_uploads_post": {
        "properties": {
          "canonical_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Canonical Url",
            "type": "string"
          },
          "doc_no": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc No"
          },
          "doc_type": {
            "default": "announcement",
            "pattern": "^(law|regulation|announcement|interpretation|guide|faq)$",
            "title": "Doc Type",
            "type": "string"
          },
          "effective_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective End"
          },
          "effective_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Start"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "issuing_authority": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Issuing Authority",
            "type": "string"
          },
          "publish_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Date"
          },
          "region_code": {
            "pattern": "^\\d{6}$",
            "title": "Region Code",
            "type": "string"
          },
          "source_level": {
            "default": "A",
            "pattern": "^[A-D]$",
            "title": "Source Level",
            "type": "string"
          },
          "source_site_id": {
            "pattern": "^[0-9a-f-]{36}$",
            "title": "Source Site Id",
            "type": "string"
          },
          "title": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_site_id",
          "title",
          "issuing_authority",
          "region_code",
          "canonical_url",
          "file"
        ],
        "title": "Body_upload_official_document_api_v1_knowledge_uploads_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionJobData": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "changed_count": {
            "title": "Changed Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discovered_count": {
            "title": "Discovered Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail_safe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail Safe"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "source_site_id": {
            "title": "Source Site Id",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trigger_type": {
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_site_id",
          "job_type",
          "trigger_type",
          "source_url",
          "status",
          "attempt_count",
          "discovered_count",
          "changed_count",
          "error_code",
          "error_detail_safe",
          "started_at",
          "finished_at",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionJobData",
        "type": "object"
      },
      "IngestionJobResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IngestionJobData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "IngestionJobResponse",
        "type": "object"
      },
...
      "ManualImportData": {
        "additionalProperties": false,
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "idempotent": {
            "title": "Idempotent",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "job_status": {
            "title": "Job Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "job_status",
          "document_id",
          "document_version_id",
          "chunk_count",
          "idempotent"
        ],
        "title": "ManualImportData",
        "type": "object"
      },
      "ManualImportResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ManualImportData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ManualImportResponse",
        "type": "object"
      },
...
      "RegisterSourceSiteRequest": {
        "additionalProperties": false,
        "properties": {
          "authority_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Authority Name",
            "type": "string"
          },
          "base_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "collection_method": {
            "pattern": "^(manual|whitelist_crawl|api|file_import)$",
            "title": "Collection Method",
            "type": "string"
          },
          "crawl_interval_minutes": {
            "anyOf": [
              {
                "maximum": 43200.0,
                "minimum": 60.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crawl Interval Minutes"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "region_code": {
            "pattern": "^\\d{6}$",
            "title": "Region Code",
            "type": "string"
          },
          "source_level": {
            "pattern": "^[A-D]$",
            "title": "Source Level",
            "type": "string"
          },
          "whitelist_rules": {
            "additionalProperties": true,
            "title": "Whitelist Rules",
            "type": "object"
          }
        },
        "required": [
          "name",
          "base_url",
          "source_level",
          "authority_name",
          "region_code",
          "collection_method"
        ],
        "title": "RegisterSourceSiteRequest",
        "type": "object"
      },
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SourceSiteData": {
        "properties": {
          "authority_name": {
            "title": "Authority Name",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "collection_method": {
            "title": "Collection Method",
            "type": "string"
          },
          "crawl_interval_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crawl Interval Minutes"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region_code": {
            "title": "Region Code",
            "type": "string"
          },
          "source_level": {
            "title": "Source Level",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "whitelist_rules": {
            "additionalProperties": true,
            "title": "Whitelist Rules",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "base_url",
          "domain",
          "source_level",
          "authority_name",
          "region_code",
          "collection_method",
          "whitelist_rules",
          "crawl_interval_minutes",
          "status",
          "last_checked_at",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "SourceSiteData",
        "type": "object"
      },
      "SourceSiteListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SourceSiteData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SourceSiteListResponse",
        "type": "object"
      },
      "SourceSiteResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SourceSiteData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SourceSiteResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TaxMind Pro",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/jobs/{job_id}": {
      "get": {
        "operationId": "get_ingestion_job_api_v1_knowledge_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ingestion Job",
        "tags": [
          "knowledge-sources"
        ]
      }
    },
...
    "/api/v1/knowledge/sources": {
      "get": {
        "operationId": "list_sources_api_v1_knowledge_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSiteListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sources",
        "tags": [
          "knowledge-sources"
        ]
      },
      "post": {
        "operationId": "register_source_api_v1_knowledge_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSourceSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Source",
        "tags": [
          "knowledge-sources"
        ]
      }
    },
    "/api/v1/knowledge/uploads": {
      "post": {
        "operationId": "upload_official_document_api_v1_knowledge_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_official_document_api_v1_knowledge_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Official Document",
        "tags": [
          "knowledge-uploads"
        ]
      }
    },
...
  }
}